src/genall/pyproject.toml
[project]
name = "genall"
version = "0.0.4"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "pathspec>=0.12.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
]
license = {file = "LICENSE"}
authors = [
    { name="Alex Rudolph", email="alex@voker.ai" },
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "types-pyyaml>=6.0.12.20250915",
]

[project.scripts]
genall = "genall.main:main"
